Clean up python files
