+++ .github/workflows/vorondesign-ci-v2.yml
@@ -61,7 +61,7 @@ jobs:
           echo "::group::check-stls"
           chmod +x ${{ github.workspace }}/.workflow_files/validate-stl_new.py
           for file in $(echo $(find ${{ github.workspace }} -type f -iname '*.STL') | tr '\n' ' '); do
-            python3 ${{ github.workspace }}/.workflow_files/validate-stl_new.py $file || touch failed
+            python3 ${{ github.workspace }}/.workflow_files/validate-stl.py $file || touch failed
           done
           if [ -f "failed" ]; then
             echo "Error occurred while validating STLs."
